projects:fmcadc5: Use iio_app
Refactor project to use common iio_app
