+++ projects/fmcadc5/src/devices/adi_hal/parameters.h
@@ -46,6 +46,7 @@
 #define SPI_DEVICE_ID				XPAR_SPI_0_DEVICE_ID
 #define GPIO_DEVICE_ID				XPAR_GPIO_0_DEVICE_ID
 #define UART_DEVICE_ID				XPAR_AXI_UART_DEVICE_ID
+#define UART_BAUDRATE                           115200
 
 #define ADC_DDR_BASEADDR			(XPAR_AXI_DDR_CNTRL_BASEADDR + 0x800000)
